# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VoIP Termination System API",
    "version": "1.0.0",
    "description": "API REST para el sistema de terminacion VoIP. Gestiona clientes, carriers, CDRs, tarifas, reportes, fraude y mas.",
    "contact": {
      "email": "admin@tellmetelecom.com"
    }
  },
  "servers": [
    {
      "url": "https://sw1.tellmetelecom.com/api/v1",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token"
      }
    },
...
  },
  "paths": {
...
    "/system/status": {
      "get": {
        "summary": "Estado del sistema",
        "tags": ["System"],
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Informacion del sistema"}}
      }
    },
...
    "/blacklist": {
      "get": {
        "summary": "Listar IPs bloqueadas",
        "tags": ["System"],
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Lista de IPs"}}
      },
      "post": {
        "summary": "Agregar IP a blacklist",
        "tags": ["System"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ip_address", "reason"],
                "properties": {
                  "ip_address": {"type": "string"},
                  "reason": {"type": "string"},
                  "permanent": {"type": "boolean"},
                  "duration": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {"201": {"description": "IP bloqueada"}}
      }
    }
  },
...
}